add imgproc snippets
